Sigma UI refactor (#2642)
* Refactor Sigma studio

* add files

* empty state
