+++ timesketch/frontend-ng/src/router.js
@@ -19,7 +19,6 @@ import VueRouter from 'vue-router'
 import Home from './views/Home'
 import Canvas from './views/Canvas'
 import Sketch from './views/Sketch'
-import Studio from './views/Studio'
 
 Vue.use(VueRouter)
 
@@ -53,6 +52,33 @@ const routes = [
         props: true,
       },
       {
+        path: 'sigma',
+        component: Canvas,
+        props: true,
+        children: [
+          {
+            path: 'new',
+            name: 'SigmaNewRule',
+            component: Canvas,
+            props: true,
+          },
+          {
+            path: 'edit/:ruleId',
+            name: 'SigmaEditRule',
+            component: Canvas,
+            props: true,
+          },
+
+        ]
+      },
+      {
+        path: 'sigma/edit',
+        name: 'SigmaEdit',
+        component: Canvas,
+        props: true,
+      },
+
+      {
         path: 'graph',
         name: 'Graph',
         component: Canvas,
@@ -66,12 +92,6 @@ const routes = [
       },
 
     ],
-  },
-  {
-    name: 'Studio',
-    path: '/studio/:type/:id',
-    component: Studio,
-    props: true,
   },
 ]
